+++ packages/image_picker/image_picker_for_web/example/integration_test/image_picker_for_web_test.dart
@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-@TestOn('chrome') // Uses dart:html
-
 import 'dart:convert';
 import 'dart:html' as html;
 import 'dart:typed_data';
@@ -11,12 +9,15 @@ import 'dart:typed_data';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:image_picker_for_web/image_picker_for_web.dart';
 import 'package:image_picker_platform_interface/image_picker_platform_interface.dart';
+import 'package:integration_test/integration_test.dart';
 
 final String expectedStringContents = "Hello, world!";
 final Uint8List bytes = utf8.encode(expectedStringContents) as Uint8List;
 final html.File textFile = html.File([bytes], "hello.txt");
 
 void main() {
+  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+
   // Under test...
   late ImagePickerPlugin plugin;
 
@@ -24,7 +25,7 @@ void main() {
     plugin = ImagePickerPlugin();
   });
 
-  test('Can select a file', () async {
+  testWidgets('Can select a file (Deprecated)', (WidgetTester tester) async {
     final mockInput = html.FileUploadInputElement();
 
     final overrides = ImagePickerPluginTestOverrides()
@@ -45,9 +46,30 @@ void main() {
     expect((await file).readAsBytes(), completion(isNotEmpty));
   });
 
+  testWidgets('Can select a file', (WidgetTester tester) async {
+    final mockInput = html.FileUploadInputElement();
+
+    final overrides = ImagePickerPluginTestOverrides()
+      ..createInputElement = ((_, __) => mockInput)
+      ..getFileFromInput = ((_) => textFile);
+
+    final plugin = ImagePickerPlugin(overrides: overrides);
+
+    // Init the pick file dialog...
+    final file = plugin.getFile();
+
+    // Mock the browser behavior of selecting a file...
+    mockInput.dispatchEvent(html.Event('change'));
+
+    // Now the file should be available
+    expect(file, completes);
+    // And readable
+    expect((await file).readAsBytes(), completion(isNotEmpty));
+  });
+
   // There's no good way of detecting when the user has "aborted" the selection.
 
-  test('computeCaptureAttribute', () {
+  testWidgets('computeCaptureAttribute', (WidgetTester tester) async {
     expect(
       plugin.computeCaptureAttribute(ImageSource.gallery, CameraDevice.front),
       isNull,
@@ -67,14 +89,14 @@ void main() {
   });
 
   group('createInputElement', () {
-    test('accept: any, capture: null', () {
+    testWidgets('accept: any, capture: null', (WidgetTester tester) async {
       html.Element input = plugin.createInputElement('any', null);
 
       expect(input.attributes, containsPair('accept', 'any'));
       expect(input.attributes, isNot(contains('capture')));
     });
 
-    test('accept: any, capture: something', () {
+    testWidgets('accept: any, capture: something', (WidgetTester tester) async {
       html.Element input = plugin.createInputElement('any', 'something');
 
       expect(input.attributes, containsPair('accept', 'any'));