+++ script/tool/lib/src/common/package_looping_command.dart
@@ -2,14 +2,59 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:async';
+
 import 'package:colorize/colorize.dart';
 import 'package:file/file.dart';
 import 'package:git/git.dart';
 import 'package:path/path.dart' as p;
+import 'package:platform/platform.dart';
 
 import 'core.dart';
 import 'plugin_command.dart';
 import 'process_runner.dart';
+
+/// Possible outcomes of a command run for a package.
+enum RunState {
+  /// The command succeeded for the package.
+  succeeded,
+
+  /// The command was skipped for the package.
+  skipped,
+
+  /// The command failed for the package.
+  failed,
+}
+
+/// The result of a [runForPackage] call.
+class PackageResult {
+  /// A successful result.
+  PackageResult.success() : this._(RunState.succeeded);
+
+  /// A run that was skipped as explained in [reason].
+  PackageResult.skip(String reason)
+      : this._(RunState.skipped, <String>[reason]);
+
+  /// A run that failed.
+  ///
+  /// If [errors] are provided, they will be listed in the summary, otherwise
+  /// the summary will simply show that the package failed.
+  PackageResult.fail([List<String> errors = const <String>[]])
+      : this._(RunState.failed, errors);
+
+  const PackageResult._(this.state, [this.details = const <String>[]]);
+
+  /// The state the package run completed with.
+  final RunState state;
+
+  /// Information about the result:
+  /// - For `succeeded`, this is empty.
+  /// - For `skipped`, it contains a single entry describing why the run was
+  ///   skipped.
+  /// - For `failed`, it contains zero or more specific error details to be
+  ///   shown in the summary.
+  final List<String> details;
+}
 
 /// An abstract base class for a command that iterates over a set of packages
 /// controlled by a standard set of flags, running some actions on each package,
@@ -19,8 +64,19 @@ abstract class PackageLoopingCommand extends PluginCommand {
   PackageLoopingCommand(
     Directory packagesDir, {
     ProcessRunner processRunner = const ProcessRunner(),
+    Platform platform = const LocalPlatform(),
     GitDir? gitDir,
-  }) : super(packagesDir, processRunner: processRunner, gitDir: gitDir);
+  }) : super(packagesDir,
+            processRunner: processRunner, platform: platform, gitDir: gitDir);
+
+  /// Packages that had at least one [logWarning] call.
+  final Set<Directory> _packagesWithWarnings = <Directory>{};
+
+  /// Number of warnings that happened outside of a [runForPackage] call.
+  int _otherWarningCount = 0;
+
+  /// The package currently being run by [runForPackage].
+  Directory? _currentPackage;
 
   /// Called during [run] before any calls to [runForPackage]. This provides an
   /// opportunity to fail early if the command can't be run (e.g., because the
@@ -33,11 +89,15 @@ abstract class PackageLoopingCommand extends PluginCommand {
   /// be included in the final error summary (e.g., a command that only has a
   /// single failure mode), or strings that should be listed for that package
   /// in the final summary. An empty list indicates success.
-  Future<List<String>> runForPackage(Directory package);
+  Future<PackageResult> runForPackage(Directory package);
 
   /// Called during [run] after all calls to [runForPackage]. This provides an
   /// opportunity to do any cleanup of run-level state.
   Future<void> completeRun() async {}
+
+  /// If [captureOutput], this is called just before exiting with all captured
+  /// [output].
+  Future<void> handleCapturedOutput(List<String> output) async {}
 
   /// Whether or not the output (if any) of [runForPackage] is long, or short.
   ///
@@ -69,20 +129,29 @@ abstract class PackageLoopingCommand extends PluginCommand {
   /// context.
   String get failureListFooter => 'See above for full details.';
 
+  /// If true, all printing (including the summary) will be redirected to a
+  /// buffer, and provided in a call to [handleCapturedOutput] at the end of
+  /// the run.
+  ///
+  /// Capturing output will disable any colorizing of output from this base
+  /// class.
+  bool get captureOutput => false;
+
   // ----------------------------------------
 
-  /// A convenience constant for [runForPackage] success that's more
-  /// self-documenting than the value.
-  static const List<String> success = <String>[];
-
-  /// A convenience constant for [runForPackage] failure without additional
-  /// context that's more self-documenting than the value.
-  static const List<String> failure = <String>[''];
-
-  /// Prints a message using a standard format indicating that the package was
-  /// skipped, with an explanation of why.
-  void printSkip(String reason) {
-    print(Colorize('SKIPPING: $reason')..darkGray());
+  /// Logs that a warning occurred, and prints `warningMessage` in yellow.
+  ///
+  /// Warnings are not surfaced in CI summaries, so this is only useful for
+  /// highlighting something when someone is already looking though the log
+  /// messages. DO NOT RELY on someone noticing a warning; instead, use it for
+  /// things that might be useful to someone debugging an unexpected result.
+  void logWarning(String warningMessage) {
+    print(Colorize(warningMessage)..yellow());
+    if (_currentPackage != null) {
+      _packagesWithWarnings.add(_currentPackage!);
+    } else {
+      ++_otherWarningCount;
+    }
   }
 
   /// Returns the identifying name to use for [package].
@@ -92,8 +161,8 @@ abstract class PackageLoopingCommand extends PluginCommand {
   /// an exact format (e.g., published name, or basename) is required, that
   /// should be used instead.
   String getPackageDescription(Directory package) {
-    String packageName = p.relative(package.path, from: packagesDir.path);
-    final List<String> components = p.split(packageName);
+    String packageName = getRelativePosixPath(package, from: packagesDir);
+    final List<String> components = p.posix.split(packageName);
     // For the common federated plugin pattern of `foo/foo_subpackage`, drop
     // the first part since it's not useful.
     if (components.length == 2 &&
@@ -103,6 +172,16 @@ abstract class PackageLoopingCommand extends PluginCommand {
     return packageName;
   }
 
+  /// Returns the relative path from [from] to [entity] in Posix style.
+  ///
+  /// This should be used when, for example, printing package-relative paths in
+  /// status or error messages.
+  String getRelativePosixPath(
+    FileSystemEntity entity, {
+    required Directory from,
+  }) =>
+      p.posix.joinAll(path.split(path.relative(entity.path, from: from.path)));
+
   /// The suggested indentation for printed output.
   String get indentation => hasLongOutput ? '' : '  ';
 
@@ -110,41 +189,75 @@ abstract class PackageLoopingCommand extends PluginCommand {
 
   @override
   Future<void> run() async {
+    bool succeeded;
+    if (captureOutput) {
+      final List<String> output = <String>[];
+      final ZoneSpecification logSwitchSpecification = ZoneSpecification(
+          print: (Zone self, ZoneDelegate parent, Zone zone, String message) {
+        output.add(message);
+      });
+      succeeded = await runZoned<Future<bool>>(_runInternal,
+          zoneSpecification: logSwitchSpecification);
+      await handleCapturedOutput(output);
+    } else {
+      succeeded = await _runInternal();
+    }
+
+    if (!succeeded) {
+      throw ToolExit(exitCommandFoundErrors);
+    }
+  }
+
+  Future<bool> _runInternal() async {
+    _packagesWithWarnings.clear();
+    _otherWarningCount = 0;
+    _currentPackage = null;
+
     await initializeRun();
 
     final List<Directory> packages = includeSubpackages
         ? await getPackages().toList()
         : await getPlugins().toList();
 
-    final Map<Directory, List<String>> results = <Directory, List<String>>{};
+    final Map<Directory, PackageResult> results = <Directory, PackageResult>{};
     for (final Directory package in packages) {
+      _currentPackage = package;
       _printPackageHeading(package);
-      results[package] = await runForPackage(package);
+      final PackageResult result = await runForPackage(package);
+      if (result.state == RunState.skipped) {
+        final String message =
+            '${indentation}SKIPPING: ${result.details.first}';
+        captureOutput ? print(message) : print(Colorize(message)..darkGray());
+      }
+      results[package] = result;
     }
+    _currentPackage = null;
 
     completeRun();
 
+    print('\n');
     // If there were any errors reported, summarize them and exit.
-    if (results.values.any((List<String> failures) => failures.isNotEmpty)) {
-      const String indentation = '  ';
-      printError(failureListHeader);
-      for (final Directory package in packages) {
-        final List<String> errors = results[package]!;
-        if (errors.isNotEmpty) {
-          final String errorIndentation = indentation * 2;
-          String errorDetails = errors.join('\n$errorIndentation');
-          if (errorDetails.isNotEmpty) {
-            errorDetails = ':\n$errorIndentation$errorDetails';
-          }
-          printError(
-              '$indentation${getPackageDescription(package)}$errorDetails');
-        }
-      }
-      printError(failureListFooter);
-      throw ToolExit(exitCommandFoundErrors);
+    if (results.values
+        .any((PackageResult result) => result.state == RunState.failed)) {
+      _printFailureSummary(packages, results);
+      return false;
     }
 
-    printSuccess('\n\nNo issues found!');
+    // Otherwise, print a summary of what ran for ease of auditing that all the
+    // expected tests ran.
+    _printRunSummary(packages, results);
+
+    print('\n');
+    _printSuccess('No issues found!');
+    return true;
+  }
+
+  void _printSuccess(String message) {
+    captureOutput ? print(message) : printSuccess(message);
+  }
+
+  void _printError(String message) {
+    captureOutput ? print(message) : printError(message);
   }
 
   /// Prints the status message indicating that the command is being run for
@@ -165,6 +278,90 @@ abstract class PackageLoopingCommand extends PluginCommand {
     } else {
       heading = '$heading...';
     }
-    print(Colorize(heading)..cyan());
+    captureOutput ? print(heading) : print(Colorize(heading)..cyan());
+  }
+
+  /// Prints a summary of packges run, packages skipped, and warnings.
+  void _printRunSummary(
+      List<Directory> packages, Map<Directory, PackageResult> results) {
+    final Set<Directory> skippedPackages = results.entries
+        .where((MapEntry<Directory, PackageResult> entry) =>
+            entry.value.state == RunState.skipped)
+        .map((MapEntry<Directory, PackageResult> entry) => entry.key)
+        .toSet();
+    final int skipCount = skippedPackages.length;
+    // Split the warnings into those from packages that ran, and those that
+    // were skipped.
+    final Set<Directory> _skippedPackagesWithWarnings =
+        _packagesWithWarnings.intersection(skippedPackages);
+    final int skippedWarningCount = _skippedPackagesWithWarnings.length;
+    final int runWarningCount =
+        _packagesWithWarnings.length - skippedWarningCount;
+
+    final String runWarningSummary =
+        runWarningCount > 0 ? ' ($runWarningCount with warnings)' : '';
+    final String skippedWarningSummary =
+        runWarningCount > 0 ? ' ($skippedWarningCount with warnings)' : '';
+    print('------------------------------------------------------------');
+    if (hasLongOutput) {
+      _printPerPackageRunOverview(packages, skipped: skippedPackages);
+    }
+    print(
+        'Ran for ${packages.length - skipCount} package(s)$runWarningSummary');
+    if (skipCount > 0) {
+      print('Skipped $skipCount package(s)$skippedWarningSummary');
+    }
+    if (_otherWarningCount > 0) {
+      print('$_otherWarningCount warnings not associated with a package');
+    }
+  }
+
+  /// Prints a one-line-per-package overview of the run results for each
+  /// package.
+  void _printPerPackageRunOverview(List<Directory> packages,
+      {required Set<Directory> skipped}) {
+    print('Run overview:');
+    for (final Directory package in packages) {
+      final bool hadWarning = _packagesWithWarnings.contains(package);
+      Styles style;
+      String summary;
+      if (skipped.contains(package)) {
+        summary = 'skipped';
+        style = hadWarning ? Styles.LIGHT_YELLOW : Styles.DARK_GRAY;
+      } else {
+        summary = 'ran';
+        style = hadWarning ? Styles.YELLOW : Styles.GREEN;
+      }
+      if (hadWarning) {
+        summary += ' (with warning)';
+      }
+
+      if (!captureOutput) {
+        summary = (Colorize(summary)..apply(style)).toString();
+      }
+      print('  ${getPackageDescription(package)} - $summary');
+    }
+    print('');
+  }
+
+  /// Prints a summary of all of the failures from [results].
+  void _printFailureSummary(
+      List<Directory> packages, Map<Directory, PackageResult> results) {
+    const String indentation = '  ';
+    _printError(failureListHeader);
+    for (final Directory package in packages) {
+      final PackageResult result = results[package]!;
+      if (result.state == RunState.failed) {
+        final String errorIndentation = indentation * 2;
+        String errorDetails = '';
+        if (result.details.isNotEmpty) {
+          errorDetails =
+              ':\n$errorIndentation${result.details.join('\n$errorIndentation')}';
+        }
+        _printError(
+            '$indentation${getPackageDescription(package)}$errorDetails');
+      }
+    }
+    _printError(failureListFooter);
   }
 }