+++ script/tool/lib/src/lint_podspecs_command.dart
@@ -14,6 +14,7 @@ import 'common/package_looping_command.dart';
 import 'common/process_runner.dart';
 
 const int _exitUnsupportedPlatform = 2;
+const int _exitPodNotInstalled = 3;
 
 /// Lint the CocoaPod podspecs and run unit tests.
 ///
@@ -24,8 +25,7 @@ class LintPodspecsCommand extends PackageLoopingCommand {
     Directory packagesDir, {
     ProcessRunner processRunner = const ProcessRunner(),
     Platform platform = const LocalPlatform(),
-  })  : _platform = platform,
-        super(packagesDir, processRunner: processRunner) {
+  }) : super(packagesDir, processRunner: processRunner, platform: platform) {
     argParser.addMultiOption('ignore-warnings',
         help:
             'Do not pass --allow-warnings flag to "pod lib lint" for podspecs '
@@ -44,44 +44,52 @@ class LintPodspecsCommand extends PackageLoopingCommand {
       'Runs "pod lib lint" on all iOS and macOS plugin podspecs.\n\n'
       'This command requires "pod" and "flutter" to be in your path. Runs on macOS only.';
 
-  final Platform _platform;
-
   @override
   Future<void> initializeRun() async {
-    if (!_platform.isMacOS) {
+    if (!platform.isMacOS) {
       printError('This command is only supported on macOS');
       throw ToolExit(_exitUnsupportedPlatform);
     }
 
-    await processRunner.run(
+    final ProcessResult result = await processRunner.run(
       'which',
       <String>['pod'],
       workingDir: packagesDir,
-      exitOnError: true,
       logOnError: true,
     );
+    if (result.exitCode != 0) {
+      printError('Unable to find "pod". Make sure it is in your path.');
+      throw ToolExit(_exitPodNotInstalled);
+    }
   }
 
   @override
-  Future<List<String>> runForPackage(Directory package) async {
+  Future<PackageResult> runForPackage(Directory package) async {
     final List<String> errors = <String>[];
-    for (final File podspec in await _podspecsToLint(package)) {
+
+    final List<File> podspecs = await _podspecsToLint(package);
+    if (podspecs.isEmpty) {
+      return PackageResult.skip('No podspecs.');
+    }
+
+    for (final File podspec in podspecs) {
       if (!await _lintPodspec(podspec)) {
         errors.add(p.basename(podspec.path));
       }
     }
-    return errors;
+    return errors.isEmpty
+        ? PackageResult.success()
+        : PackageResult.fail(errors);
   }
 
   Future<List<File>> _podspecsToLint(Directory package) async {
     final List<File> podspecs =
         await getFilesForPackage(package).where((File entity) {
       final String filePath = entity.path;
-      return p.extension(filePath) == '.podspec';
+      return path.extension(filePath) == '.podspec';
     }).toList();
 
-    podspecs.sort(
-        (File a, File b) => p.basename(a.path).compareTo(p.basename(b.path)));
+    podspecs.sort((File a, File b) => a.basename.compareTo(b.basename));
     return podspecs;
   }