+++ script/tool/test/java_test_command_test.dart
@@ -2,13 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:io' as io;
+
 import 'package:args/command_runner.dart';
 import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_utils.dart';
 import 'package:flutter_plugin_tools/src/java_test_command.dart';
-import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
 import 'mocks.dart';
@@ -17,16 +18,21 @@ import 'util.dart';
 void main() {
   group('$JavaTestCommand', () {
     late FileSystem fileSystem;
+    late MockPlatform mockPlatform;
     late Directory packagesDir;
     late CommandRunner<void> runner;
     late RecordingProcessRunner processRunner;
 
     setUp(() {
       fileSystem = MemoryFileSystem();
+      mockPlatform = MockPlatform();
       packagesDir = createPackagesDirectory(fileSystem: fileSystem);
       processRunner = RecordingProcessRunner();
-      final JavaTestCommand command =
-          JavaTestCommand(packagesDir, processRunner: processRunner);
+      final JavaTestCommand command = JavaTestCommand(
+        packagesDir,
+        processRunner: processRunner,
+        platform: mockPlatform,
+      );
 
       runner =
           CommandRunner<void>('java_test_test', 'Test for $JavaTestCommand');
@@ -48,13 +54,16 @@ void main() {
 
       await runCapturingPrint(runner, <String>['java-test']);
 
+      final Directory androidFolder =
+          plugin.childDirectory('example').childDirectory('android');
+
       expect(
         processRunner.recordedCalls,
         orderedEquals(<ProcessCall>[
           ProcessCall(
-            p.join(plugin.path, 'example/android/gradlew'),
+            androidFolder.childFile('gradlew').path,
             const <String>['testDebugUnitTest', '--info'],
-            p.join(plugin.path, 'example/android'),
+            androidFolder.path,
           ),
         ]),
       );
@@ -75,13 +84,16 @@ void main() {
 
       await runCapturingPrint(runner, <String>['java-test']);
 
+      final Directory androidFolder =
+          plugin.childDirectory('example').childDirectory('android');
+
       expect(
         processRunner.recordedCalls,
         orderedEquals(<ProcessCall>[
           ProcessCall(
-            p.join(plugin.path, 'example/android/gradlew'),
+            androidFolder.childFile('gradlew').path,
             const <String>['testDebugUnitTest', '--info'],
-            p.join(plugin.path, 'example/android'),
+            androidFolder.path,
           ),
         ]),
       );
@@ -118,7 +130,7 @@ void main() {
     });
 
     test('fails when a test fails', () async {
-      createFakePlugin(
+      final Directory pluginDir = createFakePlugin(
         'plugin1',
         packagesDir,
         platformSupport: <String, PlatformSupport>{
@@ -130,10 +142,14 @@ void main() {
         ],
       );
 
-      // Simulate failure from `gradlew`.
-      final MockProcess mockDriveProcess = MockProcess();
-      mockDriveProcess.exitCodeCompleter.complete(1);
-      processRunner.processToReturn = mockDriveProcess;
+      final String gradlewPath = pluginDir
+          .childDirectory('example')
+          .childDirectory('android')
+          .childFile('gradlew')
+          .path;
+      processRunner.mockProcessesForExecutable[gradlewPath] = <io.Process>[
+        MockProcess.failing()
+      ];
 
       Error? commandError;
       final List<String> output = await runCapturingPrint(
@@ -149,6 +165,22 @@ void main() {
           contains('plugin1:\n'
               '    example tests failed.')
         ]),
+      );
+    });
+
+    test('Skips when running no tests', () async {
+      createFakePlugin(
+        'plugin1',
+        packagesDir,
+      );
+
+      final List<String> output =
+          await runCapturingPrint(runner, <String>['java-test']);
+
+      expect(
+        output,
+        containsAllInOrder(
+            <Matcher>[contains('SKIPPING: No Java unit tests.')]),
       );
     });
   });